QEDbase

Documentation for QEDbase. This library provides (more or less) simple implementations of Lorentz vectors, Dirac spinors and Dirac matrices. These are usually used in order to do calculations in particle physics, e.g. they are part of the set of Feynman rules for a given quantum field theory (see [1]).

QEDbase.AbstractCoordinateSystem — Type

AbstractCoordinateSystem

TBW

QEDbase.AbstractDefinitePolarization — Type

Abstract base type for definite polarization of particles with is_boson.

Concrete types are PolarizationX and PolarizationY.

QEDbase.AbstractDefiniteSpin — Type

Abstract base type for definite spins of particles with is_fermion.

Concrete types are SpinUp and SpinDown.

QEDbase.AbstractDiracMatrix — Type

abstract type AbstractDiracMatrix{T} <: StaticArraysCore.FieldMatrix{4, 4, T}

Abstract type for Dirac matrices, i.e. matrix representations for linear mappings from a spinor space into another.

QEDbase.AbstractDiracVector — Type

abstract type AbstractDiracVector{T} <: StaticArraysCore.FieldVector{4, T}

Abstract type for Dirac vectors, e.g. four dimensional vectors from a spinor space.

QEDbase.AbstractFourMomentum — Type

AbstractFourMomentum

Abstract base type for four-momentas, representing one energy and three spacial components.

Also see: QEDcore.SFourMomentum, QEDcore.MFourMomentum

QEDbase.AbstractFrameOfReference — Type

AbstractFrameOfReference

TBW

QEDbase.AbstractInPhaseSpacePoint — Type

AbstractInPhaseSpacePoint

A partial type specialization on AbstractPhaseSpacePoint which can be used for dispatch in functions requiring only the in channel of the phase space to exist, for example implementations of _incident_flux. No restrictions are imposed on the out-channel, which may or may not exist.

See also: AbstractOutPhaseSpacePoint

QEDbase.AbstractIndefinitePolarization — Type

Abstract base type for indefinite polarization of particles with is_boson.

One concrete type is AllPolarization.

QEDbase.AbstractIndefiniteSpin — Type

Abstract base type for indefinite spins of particles with is_fermion.

One concrete type is AllSpin.

QEDbase.AbstractInvalidInputException — Type

Abstract base type for exceptions indicating invalid input. See InvalidInputError for a simple concrete implementation. Concrete implementations should at least implement

Base.showerror(io::IO, err::CustomInvalidError) where {CustomInvalidError<:AbstractInvalidInputException}

QEDbase.AbstractLorentzVector — Type

abstract type AbstractLorentzVector{T} <: StaticArraysCore.FieldVector{4, T}

Abstract type to model generic Lorentz vectors, i.e. elements of a minkowski-like space, where the component space is arbitray.

QEDbase.AbstractModelDefinition — Type

Abstract base type for all compute model definitions in the context of scattering processes. Every subtype of AbstractModelDefinition is associated with a fundamental interaction. Therefore, one needs to implement the following soft interface function

fundamental_interaction_type(::AbstractModelDefinition)

QEDbase.AbstractOutPhaseSpacePoint — Type

AbstractOutPhaseSpacePoint

A partial type specialization on AbstractPhaseSpacePoint which can be used for dispatch in functions requiring only the out channel of the phase space to exist. No restrictions are imposed on the in-channel, which may or may not exist.

See also: AbstractInPhaseSpacePoint

QEDbase.AbstractParticle — Type

Abstract base type for every type which might be considered a particle in the context of QED.jl. For every (concrete) subtype of AbstractParticle, there are two kinds of interface functions implemented: static functions and property functions. The static functions provide information on what kind of particle it is (defaults are written in square brackets)

    is_fermion(::AbstractParticle)::Bool [= false]
    is_boson(::AbstractParticle)::Bool [= false]
    is_particle(::AbstractParticle)::Bool [= true]
    is_anti_particle(::AbstractParticle)::Bool [= false]

If the output of those functions differ from the defaults for a subtype of AbstractParticle, these functions need to be overwritten. The second type of functions define a hard interface for AbstractParticle:

    mass(::AbstractParticle)::Real
    charge(::AbstractParticle)::Real

These functions must be implemented in order to have the subtype of AbstractParticle work with the functionalities of QEDprocesses.jl.

QEDbase.AbstractParticleStateful — Type

AbstractParticleStateful <: QEDbase.AbstractParticle

Abstract base type for the representation of a particle with a state. It requires the following interface functions to be provided:

particle_direction: Returning the particle's direction.

particle_species: Returning the particle's species.

momentum: Returning the particle's momentum.

Implementations for is_fermion, is_boson, is_particle, is_anti_particle, is_incoming, is_outgoing, mass, and charge are automatically provided using the interface functions above to fulfill the QEDbase.AbstractParticle interface.

QEDbase.AbstractParticleType — Type

AbstractParticleType <: AbstractParticle

This is the abstract base type for every species of particles. All functionalities defined on subtypes of AbstractParticleType should be static, i.e. known at compile time. For adding runtime information, e.g. four-momenta or particle states, to a particle, consider implementing a concrete subtype of AbstractParticle instead, which may have a type parameter P<:AbstractParticleType.

Concrete built-in subtypes of AbstractParticleType are available in QEDcore.jl and should always be singletons..

QEDbase.AbstractPhaseSpacePoint — Type

AbstractPhaseSpacePoint{PROC, MODEL, PSDEF, IN_PARTICLES, OUT_PARTICLES}

Representation of a point in the phase space of a process. It has several template arguments:

PROC <:AbstractProcessDefinition

MODEL <:AbstractModelDefinition

PSDEF <:AbstractPhasespaceDefinition

IN_PARTICLES <:Tuple{Vararg{AbstractParticleStateful}}: The tuple type of all the incoming [AbstractParticleStateful`](@ref)s.

OUT_PARTICLES <:Tuple{Vararg{AbstractParticleStateful}}: The tuple type of all the outgoing [AbstractParticleStateful`](@ref)s.

The following interface functions must be provided:

Base.getindex(psp::AbstractPhaseSpacePoint, dir::ParticleDirection, n::Int): Return the nth AbstractParticleStateful of the given direction. Throw BoundsError for invalid indices.

particles(psp::AbstractPhaseSpacePoint, dir::ParticleDirection): Return the particle tuple (type IN_PARTICLES or OUT_PARTICLES depending on dir)

process: Return the process.

model: Return the model.

phase_space_definition: Return the phase space definition.

From this, the following functions are automatically derived:

momentum(psp::AbstractPhaseSpacePoint, dir::ParticleDirection, n::Int): Return the momentum of the nth AbstractParticleStateful of the given direction.

momenta(psp::PhaseSpacePoint, ::ParticleDirection): Return a Tuple of all the momenta of the given direction.

Furthermore, an implementation of an AbstractPhaseSpacePoint has to verify on construction that it is valid, i.e., the following conditions are fulfilled:

IN_PARTICLES must match incoming_particles(::PROC) in length, order, and type or be an empty Tuple.

OUT_PARTICLES must match the outgoing_particles(::PROC) in length, order, and type, or be an empty Tuple.

IN_PARTICLES and OUT_PARTICLES may not both be empty.

If IN_PARTICLES is non-empty, AbstractPhaseSpacePoint <: AbstractInPhaseSpacePoint is true. Likewise, if OUT_PARTICLES is non-empty, AbstractPhaseSpacePoint <: AbstractOutPhaseSpacePoint is true. Consequently, if both IN_PARTICLES and OUT_PARTICLES are non-empty, both <: statements are true.

QEDbase.AbstractPhasespaceDefinition — Type

AbstractPhasespaceDefinition

TBW

QEDbase.AbstractPolarization — Type

Abstract base type for the polarization of particles with is_boson.

QEDbase.AbstractProcessDefinition — Type

Abstract base type for definitions of scattering processes. It is the root type for the process interface, which assumes that every subtype of AbstractProcessDefinition implements at least

incoming_particles(proc_def::AbstractProcessDefinition)
outgoing_particles(proc_def::AbstractProcessDefinition)

which return a tuple of the incoming and outgoing particles, respectively.

An AbstractProcessDefinition is also expected to contain spin and polarization information of its particles. For this, the functions

incoming_spin_pols(proc_def::AbstractProcessDefinition)
outgoing_spin_pols(proc_def::AbstractProcessDefinition)

can be overloaded. They must return a tuple of [AbstractSpinOrPolarization], where the order must match the order of the process' particles. A default implementation is provided which assumes AllSpin for every is_fermion particle and AllPolarization for every is_boson particle.

!!! note Performance It is very beneficial for the performance of derived functions if these functions return compile-time-known values.

On top of these spin and polarization functions, the following functions are automatically defined:

multiplicity(proc_def::AbstractProcessDefinition)
incoming_multiplicity(proc_def::AbstractProcessDefinition)
outgoing_multiplicity(proc_def::AbstractProcessDefinition)

Which return the number of spin and polarization combinations that should be considered for the process. For more detail, refer to the functions' documentations.

Furthermore, to calculate scattering probabilities and differential cross sections, the following interface functions need to be implemented for every combination of CustomProcess<:AbstractProcessDefinition, CustomModel<:AbstractModelDefinition, and CustomPhasespaceDefinition<:AbstractPhasespaceDefinition.

    _incident_flux(psp::InPhaseSpacePoint{CustomProcess,CustomModel})

    _matrix_element(psp::PhaseSpacePoint{CustomProcess,CustomModel})

    _averaging_norm(proc::CustomProcess)

    _is_in_phasespace(psp::PhaseSpacePoint{CustomProcess,CustomModel})

    _phase_space_factor(psp::PhaseSpacePoint{CustomProcess,CustomModel,CustomPhasespaceDefinition})

Optional is the implementation of

    _total_probability(psp::PhaseSpacePoint{CustomProcess,CustomModel,CustomPhasespaceDefinition})

to enable the calculation of total probabilities and cross sections.

QEDbase.AbstractSpin — Type

Abstract base type for the spin of particles with is_fermion.

QEDbase.AbstractSpinOrPolarization — Type

Abstract base type for the spin or polarization of particles with is_fermion or is_boson, respectively.

QEDbase.AllPolarization — Type

Concrete type indicating that a particle with is_boson has an indefinite polarization and the differential cross section calculation should average or sum over all polarizations, depending on the direction (Incoming or Outgoing) of the particle in question.

julia> using QEDbase

julia> AllPol()
all polarizations

Alias

There is a built-in alias for AllPolarization:

julia> using QEDbase

julia> AllPolarization === AllPol
true

QEDbase.AllSpin — Type

Concrete type indicating that a particle with is_fermion has an indefinite spin and the differential cross section calculation should average or sum over all spins, depending on the direction (Incoming or Outgoing) of the particle in question.

julia> using QEDbase

julia> AllSpin()
all spins

QEDbase.Incoming — Type

Incoming <: ParticleDirection

Concrete implementation of a ParticleDirection to indicate that a particle is incoming in the context of a given process. Mostly used for dispatch.

julia> using QEDbase

julia> Incoming()
incoming

ParticleDirection Interface

Besides being a subtype of ParticleDirection, Incoming has

is_incoming(::Incoming) = true
is_outgoing(::Incoming) = false

QEDbase.InvalidInputError — Type

InvalidInputError(msg::String)

Exception which is thrown if a function input is invalid.

QEDbase.Outgoing — Type

Outgoing <: ParticleDirection

Concrete implementation of a ParticleDirection to indicate that a particle is outgoing in the context of a given process. Mostly used for dispatch.

julia> using QEDbase

julia> Outgoing()
outgoing

ParticleDirection Interface

Besides being a subtype of ParticleDirection, Outgoing has

is_incoming(::Outgoing) = false
is_outgoing(::Outgoing) = true

QEDbase.ParticleDirection — Type

Abstract base type for the directions of particles in the context of processes, i.e. either they are incoming or outgoing. Subtypes of this are mostly used for dispatch.

QEDbase.PolarizationX — Type

Concrete type which indicates, that a particle with is_boson has polarization in $x$-direction.

julia> using QEDbase

julia> PolX()
x-polarized

Coordinate axes

The notion of axes, e.g. $x$- and $y$-direction is just to distinguish two orthogonal polarization directions. However, if the three-momentum of the particle with is_boson is aligned to the $z$-axis of a coordinate system, the polarization axes define the $x$- or $y$-axis, respectively.

Alias

There is a built-in alias for PolarizationX:

julia> using QEDbase

julia> PolarizationX === PolX
true

QEDbase.PolarizationY — Type

Concrete type which indicates, that a particle with is_boson has polarization in $y$-direction.

julia> using QEDbase

julia> PolY()
y-polarized

Coordinate axes

The notion of axes, e.g. $x$- and $y$-direction is just to distinguish two orthogonal polarization directions. However, if the three-momentum of the particle with is_boson is aligned to the $z$-axis of a coordinate system, the polarization axes define the $x$- or $y$-axis, respectively.

Alias

There is a built-in alias for PolarizationY:

julia> using QEDbase

julia> PolarizationY === PolY
true

QEDbase.RegistryError — Type

struct RegistryError <: Exception

Exception raised, if a certain type target_type can not be registed for a certain interface, since there needs the function func to be impleemnted.

Fields

func::Function

target_type::Any

QEDbase.SpinDown — Type

Concrete type indicating that a particle with is_fermion has spin-down.

julia> using QEDbase

julia> SpinDown()
spin down

QEDbase.SpinUp — Type

Concrete type indicating that a particle with is_fermion has spin-up.

julia> using QEDbase

julia> SpinUp()
spin up

QEDbase.SyncedPolarization — Type

SyncedPolarization{N::Int} <: AbstractIndefinitePolarization

An indefinite polarization type, indicating that multiple particles have a synced polarization. Two polarizations are considered synced when they have the same value for N. This means that the resulting multiplicity will be 2 total for all particles with the same SyncedPolarization.

Having a single SyncedPolarization{N} in a process is legal. In this case, it behaves just like an AllPolarization would.

See also: multiplicity

QEDbase.SyncedSpin — Type

SyncedSpin{N::Int} <: AbstractIndefiniteSpin

An indefinite spin type, indicating that multiple particles have a synced spin. Two spins are considered synced when they have the same value for N. This means that the resulting multiplicity will be 2 total for all particles with the same SyncedSpin.

Having a single SyncedSpin{N} in a process is legal. In this case, it behaves just like an AllSpin would.

See also: multiplicity

QEDbase.UnknownDirection — Type

UnknownDirection <: ParticleDirection

Concrete implementation of a ParticleDirection to indicate that a particle has an unknown direction. This can mean that a specific direction does not make sense in the given context, that the direction is unavailable, or that it is unnecessary.

julia> using QEDbase

julia> UnknownDirection()
unknown direction

ParticleDirection Interface

Besides being a subtype of ParticleDirection, UnknownDirection has

is_incoming(::UnknownDirection) = false
is_outgoing(::UnknownDirection) = false

QEDbase._as_svec — Function

_as_svec(x)

Accepts a single object, an SVector of objects or a tuple of objects, and returns them in a single "layer" of SVector.

Useful with base_state.

QEDbase._averaging_norm — Function

_averaging_norm(proc::AbstractProcessDefinition)

Interface function, which returns a normalization for the averaging of the squared matrix elements over spins and polarizations.

QEDbase._generate_incoming_momenta — Function

_generate_incoming_momenta(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
    phase_space_def::AbstractPhasespaceDefinition,
    in_phase_space::NTuple{N,T},
) where {N,T<:Real}

Interface function to generate the four-momenta of the incoming particles from coordinates for a given phase-space definition.

QEDbase._generate_momenta — Method

_generate_momenta(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
    phase_space_def::AbstractPhasespaceDefinition,
    in_phase_space::NTuple{N,T},
    out_phase_space::NTuple{M,T},
) where {N,M,T<:Real}

Return four-momenta for incoming and outgoing particles for given coordinate based phase-space points.

QEDbase._generate_outgoing_momenta — Function

_generate_outgoing_momenta(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
    phase_space_def::AbstractPhasespaceDefinition,
    in_phase_space::NTuple{N,T},
    out_phase_space::NTuple{M,T},
) where {N,M,T<:Real}

Interface function to generate the four-momenta of the outgoing particles from coordinates for a given phase-space definition.

QEDbase._hasmethod_registry — Method

_hasmethod_registry(fun, _)

Wrapper around Base.hasmethod with a more meaningful error message in the context of function registration.

QEDbase._incident_flux — Function

_incident_flux(in_psp::InPhaseSpacePoint{PROC,MODEL}) where {
    PROC <: AbstractProcessDefinition,
    MODEL <: AbstractModelDefinition,
}

Interface function which returns the incident flux of the given scattering process for a given InPhaseSpacePoint.

QEDbase._is_in_phasespace — Function

_is_in_phasespace(PhaseSpacePoint{PROC,MODEL}) where {
    PROC <: AbstractProcessDefinition,
    MODEL <: AbstractModelDefinition,
}

Interface function which returns true if the combination of the given incoming and outgoing phase space is physical, i.e. all momenta are on-shell and some sort of energy-momentum conservation holds.

QEDbase._matrix_element — Function

_matrix_element(PhaseSpacePoint{PROC,MODEL}) where {
    PROC <: AbstractProcessDefinition,
    MODEL <: AbstractModelDefinition,
}

Interface function which returns a tuple of scattering matrix elements for each spin and polarization combination of proc.

QEDbase._mul — Method

_mul(
    DM::Union{AbstractDiracMatrix, AbstractDiracVector},
    L::AbstractLorentzVector
) -> Any

Product of generic Lorentz vector with a Dirac tensor from the left. Basically, the multiplication is piped to the components from the Lorentz vector.

Multiplication operator

This also overloads the * operator for this types.

QEDbase._mul — Method

_mul(
    L::AbstractLorentzVector,
    DM::Union{AbstractDiracMatrix, AbstractDiracVector}
) -> Any

Product of generic Lorentz vector with a Dirac tensor from the right. Basically, the multiplication is piped to the components from the Lorentz vector.

Multiplication operator

This also overloads the * operator for this types.

QEDbase._phase_space_factor — Function

_phase_space_factor(PhaseSpacePoint{PROC,MODEL,PSDEF}) where {
    PROC <: AbstractProcessDefinition,
    MODEL <: AbstractModelDefinition
    PSDEF <: AbstractPhasespaceDefinition,
}

Interface function, which returns the pre-differential factor of the invariant phase space intergral measure.

Convention

It is assumed, that this function returns the value of

\[\mathrm{d}\Pi_n:= \prod_{i=1}^N \frac{\mathrm{d}^3p_i}{(2\pi)^3 2 p_i^0} H(P_t, p_1, \dots, p_N),\]

where $H(\dots)$ is a characteristic function (or distribution) which constrains the phase space, e.g. $\delta^{(4)}(P_t - \sum_i p_i)$.

QEDbase._spin_index — Method

_spin_index(_::AbstractDefiniteSpin) -> Int64

Utility function, which converts subtypes of AbstractDefiniteSpin into an integer representation, e.g. an index:

SpinUp $\rightarrow 1$

SpinDown $\rightarrow 2$

This is useful in the implementation of some of base_state's overloads.

QEDbase._total_probability — Function

_total_probability(in_psp::InPhaseSpacePoint{PROC,MODEL}) where {
    PROC <: AbstractProcessDefinition,
    MODEL <: AbstractModelDefinition,
}

Interface function for the combination of a scattering process and a physical model. Return the total of a given process and model for a passed InPhaseSpacePoint.

total cross section

Given an implementation of this method and _incident_flux, the respective function for the total cross section total_cross_section is also available.

QEDbase.assert_onshell — Method

assert_onshell(mom, mass)

Assertion if a FourMomentum mom is on-shell w.r.t a given mass mass.

See also

The precision of this functions is explained in isonshell.

QEDbase.base_state — Function

    base_state(
        particle::AbstractParticleType,
        direction::ParticleDirection,
        momentum::QEDbase.AbstractFourMomentum,
        [spin_or_pol::AbstractSpinOrPolarization]
    )

Return the base state of a directed on-shell particle with a given four-momentum. For internal usage only.

Input

particle – the type of the particle, i.e., an instance of an AbstractParticleType, e.g. QEDcore.Electron, QEDcore.Positron, or QEDcore.Photon.

direction – the direction of the particle, i.e. Incoming or Outgoing.

momentum – the four-momentum of the particle

[spin_or_pol] – if given, the spin or polarization of the particle, e.g. SpinUp/SpinDown or PolarizationX/PolarizationY.

Output

The output type of base_state depends on wether the spin or polarization of the particle passed in is specified or not.

If spin_or_pol is passed, the output of base_state is

base_state(::Fermion,     ::Incoming, mom, spin_or_pol) # -> BiSpinor
base_state(::AntiFermion, ::Incoming, mom, spin_or_pol) # -> AdjointBiSpinor
base_state(::Fermion,     ::Outgoing, mom, spin_or_pol) # -> AdjointBiSpinor
base_state(::AntiFermion, ::Outgoing, mom, spin_or_pol) # -> BiSpinor
base_state(::Photon,      ::Incoming, mom, spin_or_pol) # -> SLorentzVector{ComplexF64}
base_state(::Photon,      ::Outgoing, mom, spin_or_pol) # -> SLorentzVector{ComplexF64}

If spin_or_pol is of type AllPolarization or AllSpin, the output is an SVector with both spin/polarization alignments:

base_state(::Fermion,     ::Incoming, mom) # -> SVector{2,BiSpinor}
base_state(::AntiFermion, ::Incoming, mom) # -> SVector{2,AdjointBiSpinor}
base_state(::Fermion,     ::Outgoing, mom) # -> SVector{2,AdjointBiSpinor}
base_state(::AntiFermion, ::Outgoing, mom) # -> SVector{2,BiSpinor}
base_state(::Photon,      ::Incoming, mom) # -> SVector{2,SLorentzVector{ComplexF64}}
base_state(::Photon,      ::Outgoing, mom) # -> SVector{2,SLorentzVector{ComplexF64}}

Example

using QEDbase

mass = 1.0                              # set electron mass to 1.0
px,py,pz = rand(3)                      # generate random spatial components
E = sqrt(px^2 + py^2 + pz^2 + mass^2)   # compute energy, i.e. the electron is on-shell
mom = SFourMomentum(E, px, py, pz)      # initialize the four-momentum of the electron

# compute the state of an incoming electron with spin = SpinUp
# note: base_state is not exported!
electron_state = base_state(QEDcore.Electron(), Incoming(), mom, SpinUp())

julia> using QEDbase; using QEDcore

julia> mass = 1.0; px,py,pz = (0.1, 0.2, 0.3); E = sqrt(px^2 + py^2 + pz^2 + mass^2); mom = SFourMomentum(E, px, py, pz)
4-element SFourMomentum with indices SOneTo(4):
 1.0677078252031311
 0.1
 0.2
 0.3

julia> electron_state = base_state(Electron(), Incoming(), mom, SpinUp())
4-element BiSpinor with indices SOneTo(4):
   1.4379526505428235 + 0.0im
                  0.0 + 0.0im
 -0.20862995724285552 + 0.0im
 -0.06954331908095185 - 0.1390866381619037im

julia> electron_states = base_state(Electron(), Incoming(), mom, AllSpin())
2-element StaticArraysCore.SVector{2, BiSpinor} with indices SOneTo(2):
 [1.4379526505428235 + 0.0im, 0.0 + 0.0im, -0.20862995724285552 + 0.0im, -0.06954331908095185 - 0.1390866381619037im]
 [0.0 + 0.0im, 1.4379526505428235 + 0.0im, -0.06954331908095185 + 0.1390866381619037im, 0.20862995724285552 + 0.0im]

Iterator convenience

The returned objects of base_state can be consistently wrapped in an SVector for iteration using _as_svec.

This way, a loop like the following becomes possible when spin may be definite or indefinite.

for state in QEDbase._as_svec(base_state(Electron(), Incoming(), momentum, spin))
    # ...
end

Conventions

For an incoming fermion with momentum $p$, we use the explicit formula:

\[u_\sigma(p) = \frac{\gamma^\mu p_\mu + m}{\sqrt{\vert p_0\vert + m}} \eta_\sigma,\]

where the elementary base spinors are given as

\[\eta_1 = (1, 0, 0, 0)^T\\ \eta_2 = (0, 1, 0, 0)^T\]

For an outgoing anti-fermion with momentum $p$, we use the explicit formula:

\[v_\sigma(p) = \frac{-\gamma^\mu p_\mu + m}{\sqrt{\vert p_0\vert + m}} \chi_\sigma,\]

where the elementary base spinors are given as

\[\chi_1 = (0, 0, 1, 0)^T\\ \chi_2 = (0, 0, 0, 1)^T\]

For outgoing fermions and incoming anti-fermions with momentum $p$, the base state is given as the Dirac-adjoint of the respective incoming fermion or outgoing anti-fermion state:

\[\overline{u}_\sigma(p) = u_\sigma^\dagger \gamma^0\\ \overline{v}_\sigma(p) = v_\sigma^\dagger \gamma^0\]

where $v_\sigma$ is the base state of the respective outgoing anti-fermion.

For a photon with four-momentum $k^\mu = \omega (1, \cos\phi \sin\theta, \sin\phi \sin\theta, \cos\theta)$, the two polarization vectors are given as

\[\begin{align*} \epsilon^\mu_1 &= (0, \cos\theta \cos\phi, \cos\theta \sin\phi, -\sin\theta),\\ \epsilon^\mu_2 &= (0, -\sin\phi, \cos\phi, 0). \end{align*}\]

Warning

In the current implementation there are no checks built-in, which verify the passed arguments whether they describe on-shell particles, i.e. p*p≈mass^2. Using base_state with off-shell particles will cause unpredictable behavior.

QEDbase.charge — Function

charge(::AbstractParticle)::Real

Interface function for particles. Return the electric charge of a particle (in units of the elementary electric charge).

This needs to be implemented for each concrete subtype of AbstractParticle.

QEDbase.differential_cross_section — Method

differential_cross_section(phase_space_point::PhaseSpacePoint)

If the given phase spaces are physical, return differential cross section evaluated on a phase space point. Zero otherwise.

QEDbase.differential_probability — Method

differential_probability(phase_space_point::AbstractPhaseSpacePoint)

If the given phase spaces are physical, return differential probability evaluated on a phase space point. Zero otherwise.

QEDbase.fundamental_interaction_type — Function

fundamental_interaction_type(models_def::AbstractModelDefinition)

Return the fundamental interaction associated with the passed model definition.

QEDbase.getBeta — Method

getBeta(lv)

Return magnitude of the beta vector for a given LorentzVectorLike, i.e. the magnitude of the LorentzVectorLike divided by its 0-component.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to sqrt(px^2 + py^2 + pz^2)/E.

QEDbase.getCosPhi — Method

getCosPhi(lv)

Return the cosine of the phi angle for a given LorentzVectorLike.

Note

This is an equivalent but faster version of cos(getPhi(lv)); see getPhi.

QEDbase.getCosTheta — Method

getCosTheta(lv)

Return the cosine of the theta angle for a given LorentzVectorLike.

Note

This is an equivalent but faster version of cos(getTheta(lv)); see getTheta.

QEDbase.getE — Method

getE(lv)

Return the energy component of a given LorentzVectorLike, i.e. its 0-component.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to E.

QEDbase.getEnergy — Function

Function alias for getE.

QEDbase.getGamma — Method

getGamma(lv)

Return the relativistic gamma factor for a given LorentzVectorLike, i.e. the inverse square root of one minus the beta vector squared.

Example

If (E,px,py,pz) is a LorentzVectorLike with beta vector β, this is equivalent to 1/sqrt(1- β^2).

QEDbase.getInvariantMass — Method

getInvariantMass(lv)

Return the the invariant mass of a given LorentzVectorLike, i.e. the square root of the minkowski dot with itself.

Example

If (t,x,y,z) is a LorentzVectorLike, this is equivalent to sqrt(t^2 - (x^2 + y^2 + z^2)).

Note

If the squared invariant mass m2 is negative, -sqrt(-m2) is returned.

QEDbase.getInvariantMass2 — Method

getInvariantMass2(lv)

Return the squared invariant mass of a given LorentzVectorLike, i.e. the minkowski dot with itself.

Example

If (t,x,y,z) is a LorentzVectorLike, this is equivalent to t^2 - (x^2 + y^2 + z^2).

QEDbase.getMag — Function

Functiom alias for getMagnitude.

QEDbase.getMag2 — Function

Functiom alias for getMagnitude2.

QEDbase.getMagnitude — Method

getMagnitude(lv)

Return the magnitude of a given LorentzVectorLike, i.e. the euklidian norm spatial components.

Example

If (t,x,y,z) is a LorentzVectorLike, this is equivalent to sqrt(x^2 + y^2 + z^2).

Warning

This function differs from a similar function for the TLorentzVector used in the famous ROOT library.

QEDbase.getMagnitude2 — Method

getMagnitude2(lv)

Return the square of the magnitude of a given LorentzVectorLike, i.e. the sum of the squared spatial components.

Example

If (t,x,y,z) is a LorentzVectorLike, this is equivalent to x^2+ y^2 + z^2.

Warning

This function differs from a similar function for the TLorentzVector used in the famous ROOT library.

QEDbase.getMass — Function

Function alias for getInvariantMass.

QEDbase.getMass2 — Function

Function alias for getInvariantMass2

QEDbase.getMinus — Method

getMinus(lv)

Return the minus component for a given LorentzVectorLike in light-cone coordinates.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to (E-pz)/2.

Note

The light-cone coordinates are defined w.r.t. to the 3-axis.

Warning

The definition `p^- := (E - p_z)/2 differs from the usual definition of light-cone coordinates in general relativity.

QEDbase.getMt — Function

Function alias for getTransverseMass

QEDbase.getMt2 — Function

Function alias for getTransverseMass2

QEDbase.getPerp — Function

Function alias for getTransverseMomentum.

QEDbase.getPerp2 — Function

Function alias for getTransverseMomentum2.

QEDbase.getPhi — Method

getPhi(lv)

Return the phi angle for a given LorentzVectorLike, i.e. the azimuthal angle of its spatial components in spherical coordinates.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to atan(py,px).

Note

The spherical coordinates are defined w.r.t. to the 3-axis.

QEDbase.getPlus — Method

getPlus(lv)

Return the plus component for a given LorentzVectorLike in light-cone coordinates.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to (E+pz)/2.

Note

The light-cone coordinates are defined w.r.t. to the 3-axis.

Warning

The definition `p^+ := (E + p_z)/2 differs from the usual definition of light-cone coordinates in general relativity.

QEDbase.getPt — Function

Function alias for getTransverseMomentum.

QEDbase.getPt2 — Function

Function alias for getTransverseMomentum2.

QEDbase.getPx — Method

getPx(lv)

Return the $p_x$ component of a given LorentzVectorLike, i.e. its 1-component.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to px.

QEDbase.getPy — Method

getPy(lv)

Return the $p_y$ component of a given LorentzVectorLike, i.e. its 2-component.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to py.

QEDbase.getPz — Method

getPz(lv)

Return the $p_z$ component of a given LorentzVectorLike, i.e. its 3-component.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to pz.

QEDbase.getRapidity — Method

getRapidity(lv)

Return the rapidity for a given LorentzVectorLike.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to 0.5*log((E+pz)/(E-pz)).

Note

The transverse components are defined w.r.t. to the 3-axis.

QEDbase.getRho — Function

Function alias for getMagnitude

QEDbase.getRho2 — Function

Function alias for getMagnitude2

QEDbase.getSinPhi — Method

getSinPhi(lv)

Return the sine of the phi angle for a given LorentzVectorLike.

Note

This is an equivalent but faster version of sin(getPhi(lv)); see getPhi.

QEDbase.getT — Function

getT(lv)

Return the 0-component of a given LorentzVectorLike.

Example

If (t,x,y,z) is a LorentzVectorLike, this is equivalent to t.

QEDbase.getTheta — Method

getTheta(lv)

Return the theta angle for a given LorentzVectorLike, i.e. the polar angle of its spatial components in spherical coordinates.

Example

If (E,px,py,pz) is a LorentzVectorLike with magnitude rho, this is equivalent to arccos(pz/rho), which is also equivalent to arctan(sqrt(px^2+py^2)/pz).

Note

The spherical coordinates are defined w.r.t. to the 3-axis.

QEDbase.getTransverseMass — Method

getTransverseMass(lv)

Return the transverse momentum for a given LorentzVectorLike, i.e. the square root of its squared transverse mass.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to sqrt(E^2 - pz^2).

Note

The transverse components are defined w.r.t. to the 3-axis.

Note

If the squared transverse mass mT2 is negative, -sqrt(-mT2) is returned.

QEDbase.getTransverseMass2 — Method

getTransverseMass2(lv)

Return the squared transverse mass for a given LorentzVectorLike, i.e. the difference of its squared 0- and 3-component.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to E^2 - pz^2.

Note

The transverse components are defined w.r.t. to the 3-axis.

QEDbase.getTransverseMomentum — Method

getTransverseMomentum(lv)

Return the transverse momentum for a given LorentzVectorLike, i.e. the magnitude of its transverse components.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to sqrt(px^2 + py^2).

Note

The transverse components are defined w.r.t. to the 3-axis.

QEDbase.getTransverseMomentum2 — Method

getTransverseMomentum2(lv)

Return the squared transverse momentum for a given LorentzVectorLike, i.e. the sum of its squared 1- and 2-component.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to px^2 + py^2.

Note

The transverse components are defined w.r.t. to the 3-axis.

QEDbase.getX — Function

getX(lv)

Return the 1-component of a given LorentzVectorLike.

Example

If (t,x,y,z) is a LorentzVectorLike, this is equivalent to x.

QEDbase.getY — Function

getY(lv)

Return the 2-component of a given LorentzVectorLike.

Example

If (t,x,y,z) is a LorentzVectorLike, this is equivalent to y.

QEDbase.getZ — Function

getZ(lv)

Return the 3-component of a given LorentzVectorLike.

Example

If (t,x,y,z) is a LorentzVectorLike, this is equivalent to z.

QEDbase.in_phase_space_dimension — Function

in_phase_space_dimension(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
)

TBW

QEDbase.incoming_multiplicity — Method

incoming_multiplicity(proc::AbstractProcessDefinition)

Return the number of spin and polarization combinations represented by procs incoming particles. This function only considers the incoming particles' spins and polarizations, returned by incoming_spin_pols for proc.

!!! note Incoming and Outgoing Spins/Polarizations Note that the total multiplicity is not necessarily the incoming and outgoing multiplicities multiplied. For the total process multiplicity, see multiplicity.

See also: SyncedPolarization, SyncedSpin, multiplicity, outgoing_multiplicity

QEDbase.incoming_particles — Function

incoming_particles(proc_def::AbstractProcessDefinition)

Interface function for scattering processes. Return a tuple of the incoming particles for the given process definition. This function needs to be given to implement the scattering process interface.

!!! note Performance It is very beneficial for the performance of derived functions if this function returns compile-time-known values.

See also: AbstractParticleType

QEDbase.incoming_spin_pols — Function

incoming_spin_pols(proc_def::AbstractProcessDefinition)

Interface function for scattering processes. Return the tuple of spins or polarizations for the given process definition. The order must be the same as the particles returned from incoming_particles. A default implementation is provided, returning AllSpin for every is_fermion and AllPolarization for every is_boson.

!!! note Performance It is very beneficial for the performance of derived functions if this function returns compile-time-known values.

See also: AbstractSpinOrPolarization

QEDbase.is_anti_particle — Method

is_anti_particle(_::AbstractParticle) -> Bool

Interface function for particles. Return true if the passed subtype of AbstractParticle can be considered an anti particle as distinct from their particle counterpart, and false otherwise. The default implementation of is_anti_particle for every subtype of AbstractParticle will always return false.

QEDbase.is_boson — Method

is_boson(_::AbstractParticle)

Interface function for particles. Return true if the passed subtype of AbstractParticle can be considered a boson in the sense of particle statistics, and false otherwise. The default implementation of is_boson for every subtype of AbstractParticle will always return false.

QEDbase.is_fermion — Method

is_fermion(_::AbstractParticle) -> Bool

Interface function for particles. Return true if the passed subtype of AbstractParticle can be considered a fermion in the sense of particle statistics, and false otherwise. The default implementation of is_fermion for every subtype of AbstractParticle will always return false.

QEDbase.is_incoming — Function

is_incoming(dir::ParticleDirection)
is_incoming(particle::AbstractParticleStateful)

Convenience function that returns true for Incoming and incoming AbstractParticleStateful and false otherwise.

QEDbase.is_outgoing — Function

is_outgoing(dir::ParticleDirection)
is_outgoing(particle::AbstractParticleStateful)

Convenience function that returns true for Outgoing and outgoing AbstractParticleStateful and false otherwise.

QEDbase.is_particle — Method

is_particle(_::AbstractParticle) -> Bool

Interface function for particles. Return true if the passed subtype of AbstractParticle can be considered a particle as distinct from anti-particles, and false otherwise. The default implementation of is_particle for every subtype of AbstractParticle will always return true.

QEDbase.isonshell — Method

isonshell(mom, mass)

On-shell check of a given four-momentum mom w.r.t. a given mass mass.

Precision

For AbstactFourMomentum, the element type is fixed to Float64, limiting the precision of comparisons between elements. The current implementation has been tested within the boundaries for energy scales E with 1e-9 <= E <= 1e5. In those bounds, the mass error, which is correctly detected as off-shell, is 1e-4 times the mean value of the components, but has at most the value 0.01, e.g. at the high energy end. The energy scales correspond to 0.5 meV for the lower bound and 50 GeV for the upper bound.

FourMomenta with real entries

if AbstractFourMomentum is updated to elementtypes T<:Real, the AbstractLorentzVector should be updated with the AbstractFourMomentum.

QEDbase.mass — Function

mass(particle::AbstractParticle)::Real

Interface function for particles. Return the rest mass of a particle (in units of the electron mass).

This needs to be implemented for each concrete subtype of AbstractParticle.

QEDbase.mdot — Function

Function alias for minkowski_dot.

QEDbase.minkowski_dot — Method

minkowski_dot(v1,v2)

Return the Minkowski dot product of two LorentzVectorLike.

Example

If (t1,x1,y1,z1) and (t2,x2,y2,z2) are two LorentzVectorLike, this is equivalent to

t1*t2 - (x1*x2 + y1*y2 + z1*z2)

Note

We use the mostly minus metric.

QEDbase.model — Function

model(psp::AbstractPhaseSpacePoint)

Return the phase space point's set model.

QEDbase.momenta — Method

momenta(psp::AbstractPhaseSpacePoint, ::ParticleDirection)

Return a Tuple of all the particles' momenta for the given ParticleDirection.

QEDbase.momentum — Function

momentum(part::AbstractParticleStateful)

Interface function that must return the particle's AbstractFourMomentum.

QEDbase.momentum — Method

momentum(psp::AbstractPhaseSpacePoint, dir::ParticleDirection, species::AbstractParticleType, n::Int)

Returns the momentum of the nth particle in the given AbstractPhaseSpacePoint which has direction dir and species species. If n is outside the valid range for this phase space point, a BoundsError is thrown.

Note

This function accepts n as an Int value. If n is a compile-time constant (for example, a literal 1 or 2), you can use Val(n) instead to call a zero overhead version of this function.

QEDbase.momentum — Method

momentum(psp::AbstractPhaseSpacePoint, dir::ParticleDirection, species::AbstractParticleType)

Returns the momentum of the particle in the given AbstractPhaseSpacePoint with dir and species, if there is only one such particle. If there are multiple or none, an InvalidInputError is thrown.

QEDbase.momentum — Method

momentum(psp::AbstractPhaseSpacePoint, dir::ParticleDirection, n::Int)

Returns the momentum of the nth particle in the given AbstractPhaseSpacePoint which has direction dir. If n is outside the valid range for this phase space point, a BoundsError is thrown.

QEDbase.momentum — Method

momentum(psp::AbstractPhaseSpacePoint, dir::ParticleDirection, species::AbstractParticleType, n::Val{N})

Returns the momentum of the nth particle in the given AbstractPhaseSpacePoint which has direction dir and species species. If n is outside the valid range for this phase space point, a BoundsError is thrown.

Note

This function accepts n as a Val{N} type, i.e., a compile-time constant value (for example a literal 1 or 2). This allows this function to add zero overhead, but only if N is actually known at compile time. If it is not, use the overload of this function that uses n::Int instead. That function is faster than calling this one with Val(n).

QEDbase.multiplicity — Method

multiplicity(proc::AbstractProcessDefinition)

Return the number of spin and polarization combinations represented by proc total. This depends on the specific AbstractSpinOrPolarizations returned by spin_pols for proc. For example, a default Compton process with four indefinite spins/polarizations has a multiplicity of 2^4 = 16. A Compton process with many incoming photons that have synced polarizations will still have a multiplicity of 16.

!!! note Performance As long as incoming_spin_pols and outgoing_spin_pols can be evaluated at compile time, this function is completely compiled away.

!!! note Incoming and Outgoing Spins/Polarizations Note that the total multiplicity is not necessarily the incoming and outgoing multiplicities multiplied. This is the case when incoming and outgoing particles are synced with one another.

See also: SyncedPolarization, SyncedSpin, incoming_multiplicity, outgoing_multiplicity

QEDbase.number_incoming_particles — Method

number_incoming_particles(proc_def::AbstractProcessDefinition)

Return the number of incoming particles of a given process.

QEDbase.number_outgoing_particles — Method

number_outgoing_particles(proc_def::AbstractProcessDefinition)

Return the number of outgoing particles of a given process.

QEDbase.number_particles — Method

number_particles(proc_def::AbstractProcessDefinition, particle::AbstractParticleStateful)

Return the number of particles of the given particle's direction and species in the given process definition.

QEDbase.number_particles — Method

number_particles(proc_def::AbstractProcessDefinition, dir::ParticleDirection)

Convenience function dispatching to number_incoming_particles or number_outgoing_particles depending on the given direction, returning the number of incoming or outgoing particles, respectively.

QEDbase.number_particles — Method

number_particles(proc_def::AbstractProcessDefinition, dir::ParticleDirection, species::AbstractParticleType)

Return the number of particles of the given direction and species in the given process definition.

QEDbase.out_phase_space_dimension — Function

out_phase_space_dimension(
    proc::AbstractProcessDefinition,
    model::AbstractModelDefinition,
)

TBW

QEDbase.outgoing_multiplicity — Method

outgoing_multiplicity(proc::AbstractProcessDefinition)

Return the number of spin and polarization combinations represented by procs outgoing particles. This function only considers the outgoing particles' spins and polarizations, returned by outgoing_spin_pols for proc.

!!! note Incoming and Outgoing Spins/Polarizations Note that the total multiplicity is not necessarily the incoming and outgoing multiplicities multiplied. For the total process multiplicity, see multiplicity.

See also: SyncedPolarization, SyncedSpin, multiplicity, incoming_multiplicity

QEDbase.outgoing_particles — Function

outgoing_particles(proc_def::AbstractProcessDefinition)

Interface function for scattering processes. Return the tuple of outgoing particles for the given process definition. This function needs to be given to implement the scattering process interface.

!!! note Performance It is very beneficial for the performance of derived functions if this function returns compile-time-known values.

See also: AbstractParticleType

QEDbase.outgoing_spin_pols — Function

outgoing_spin_pols(proc_def::AbstractProcessDefinition)

Interface function for scattering processes. Return the tuple of spins or polarizations for the given process definition. The order must be the same as the particles returned from outgoing_particles. A default implementation is provided, returning AllSpin for every is_fermion and AllPolarization for every is_boson.

!!! note Performance It is very beneficial for the performance of derived functions if this function returns compile-time-known values.

See also: AbstractSpinOrPolarization

QEDbase.particle_direction — Function

particle_direction(part::AbstractParticleStateful)

Interface function that must return the particle's ParticleDirection.

QEDbase.particle_species — Function

particle_species(part::AbstractParticleStateful)

Interface function that must return the particle's AbstractParticleType, e.g. QEDcore.Electron().

QEDbase.particles — Method

particles(proc_def::AbstractProcessDefinition, ::ParticleDirection)

Convenience function dispatching to incoming_particles or outgoing_particles depending on the given direction.

QEDbase.phase_space_definition — Function

phase_space_definition(psp::AbstractPhaseSpacePoint)

Return the phase space point's set phase space definition.

QEDbase.process — Function

process(psp::AbstractPhaseSpacePoint)

Return the phase space point's set process.

QEDbase.propagator — Function

propagator(particle::AbstractParticleType, mom::QEDbase.AbstractFourMomentum, [mass::Real])

Return the propagator of a particle for a given four-momentum. If mass is passed, the respective propagator for massive particles is used, if not, it is assumed the particle passed in is massless.

Convention

There are two types of implementations for propagators given in QEDProcesses: For a BosonLike particle with four-momentum $k$ and mass $m$, the propagator is given as

\[D(k) = \frac{1}{k^2 - m^2}.\]

For a FermionLike particle with four-momentum $p$ and mass $m$, the propagator is given as

\[S(p) = \frac{\gamma^\mu p_\mu + mass}{p^2 - m^2}.\]

Warning

This function does not throw when the given particle is off-shell. If an off-shell particle is passed, the function propagator returns Inf.

QEDbase.register_LorentzVectorLike — Method

register_LorentzVectorLike(T)

Function to register a custom type as a LorentzVectorLike.

Ensure the passed custom type has implemented at least the function getT, getX, getY, getZ and enables getter functions of the lorentz vector library for the given type. If additionally the functions setT!, setX!, setY!, setZ! are implemened for the passed custom type, also the setter functions of the Lorentz vector interface are enabled.

QEDbase.setCosTheta! — Method

setCosTheta!(lv,value)

Sets the cosine of the theta angle of a LorentzVectorLike to a given value.

Note

The value set with setCosTheta! is then returned by getCosTheta. Since the cosine of the theta angle is computed on the call of getCosTheta, the setter setCosTheta! changes several components of the given LorentzVectorLike.

QEDbase.setE! — Method

setE!(lv,value)

Sets the energy component of a given LorentzVectorLike to a given value.

Note

The value set with setE! is then returned by getE.

QEDbase.setEnergy! — Function

Function alias for setE!.

QEDbase.setMinus! — Method

setMinus!(lv,value)

Sets the minus component of a LorentzVectorLike to a given value.

Note

The value set with setMinus! is then returned by getMinus. Since the minus component is computed on the call of getMinus, the setter setMinus! changes several components of the given LorentzVectorLike.

QEDbase.setMt! — Function

Function alias for setTransverseMass!.

QEDbase.setPerp! — Function

Function alias for setTransverseMomentum!.

QEDbase.setPhi! — Method

setPhi!(lv,value)

Sets the phi angle of a LorentzVectorLike to a given value.

Note

The value set with setPhi! is then returned by getPhi. Since the phi angle is computed on the call of getPhi, the setter setPhi! changes several components of the given LorentzVectorLike.

QEDbase.setPlus! — Method

setPlus!(lv,value)

Sets the plus component of a LorentzVectorLike to a given value.

Note

The value set with setPlus! is then returned by getPlus. Since the plus component is computed on the call of getPlus, the setter setPlus! changes several components of the given LorentzVectorLike.

QEDbase.setPt! — Function

Function alias for setTransverseMomentum!.

QEDbase.setPx! — Method

setPx!(lv,value)

Sets the 1-component of a given LorentzVectorLike to a given value.

Note

The value set with setPx! is then returned by getPx.

QEDbase.setPy! — Method

setPy!(lv,value)

Sets the 2-component of a given LorentzVectorLike to a given value.

Note

The value set with setPy! is then returned by getPy.

QEDbase.setPz! — Method

setPz!(lv,value)

Sets the 3-component of a given LorentzVectorLike to a given value.

Note

The value set with setPz! is then returned by getPz.

QEDbase.setRapidity! — Method

setRapidity!(lv,value)

Sets the rapidity of a LorentzVectorLike to a given value.

Note

The value set with setRapidity! is then returned by getRapidity. Since the rapidity is computed on the call of setRapidity, the setter setRapidity! changes several components of the given LorentzVectorLike.

QEDbase.setRho! — Method

setRho!(lv,value)

Sets the magnitude of a LorentzVectorLike to a given value.

Note

The value set with setRho! is then returned by getRho. Since the magnitude is computed on the call of getRho, the setter setRho! changes several components of the given LorentzVectorLike.

QEDbase.setT! — Function

setT!(lv,value)

Sets the 0-component of a given LorentzVectorLike to the given value.

QEDbase.setTheta! — Method

setTheta!(lv,value)

Sets the theta angle of a LorentzVectorLike to a given value.

Note

The value set with setTheta! is then returned by getTheta. Since the theta angle is computed on the call of getTheta, the setter setTheta! changes several components of the given LorentzVectorLike.

QEDbase.setTransverseMass! — Method

setTransverseMass!(lv,value)

Sets the transverse mass of a LorentzVectorLike to a given value.

Note

The value set with setTransverseMass! is then returned by getTransverseMass. Since the transverse mass is computed on the call of getTransverseMass, the setter setTransverseMass! changes several components of the given LorentzVectorLike.

QEDbase.setTransverseMomentum! — Method

setTransverseMomentum!(lv,value)

Sets the transverse momentum of a LorentzVectorLike to a given value.

Note

The value set with setTransverseMomentum! is then returned by getTransverseMomentum. Since the transverse momentum is computed on the call of getTransverseMomentum, the setter setTransverseMomentum! changes several components of the given LorentzVectorLike.

QEDbase.setX! — Function

setX!(lv,value)

Sets the 1-component of a given LorentzVectorLike to the given value.

QEDbase.setY! — Function

setY!(lv,value)

Sets the 2-component of a given LorentzVectorLike to the given value.

QEDbase.setZ! — Function

setZ!(lv,value)

Sets the 3-component of a given LorentzVectorLike to the given value.

QEDbase.spin_pols — Method

spin_pols(proc_def::AbstractProcessDefinition, dir::ParticleDirection)

Return the tuple of spins and polarizations for the process in the given direction. Dispatches to incoming_spin_pols or outgoing_spin_pols.

QEDbase.total_cross_section — Method

total_cross_section(in_psp::AbstractInPhaseSpacePoint)

Return the total cross section for a given AbstractInPhaseSpacePoint.

QEDbase.total_probability — Method

total_probability(in_psp::AbstractInPhaseSpacePoint)

Return the total probability of a given AbstractInPhaseSpacePoint.

QEDbase.unsafe_differential_cross_section — Method

unsafe_differential_cross_section(phase_space_point::AbstractPhaseSpacePoint)

Return the differential cross section evaluated on a phase space point without checking if the given phase space is physical.

QEDbase.unsafe_differential_probability — Method

unsafe_differential_probability(phase_space_point::AbstractPhaseSpacePoint)

Return differential probability evaluated on a phase space point without checking if the given phase space(s) are physical.