QEDbase

Lorentz Vector Interface

QEDbase.RegistryError — Type

struct RegistryError <: Exception

Exception raised, if a certain type target_type can not be registed for a certain interface, since there needs the function func to be impleemnted.

Fields

func::Function

target_type::Any

QEDbase._hasmethod_registry — Method

_hasmethod_registry(fun, _)

Wrapper around Base.hasmethod with a more meaningful error message in the context of function registration.

QEDbase.getBeta — Method

getBeta(lv)

Return magnitude of the beta vector for a given LorentzVectorLike, i.e. the magnitude of the LorentzVectorLike divided by its 0-component.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to sqrt(px^2 + py^2 + pz^2)/E.

QEDbase.getCosPhi — Method

getCosPhi(lv)

Return the cosine of the phi angle for a given LorentzVectorLike.

Note

This is an equivalent but faster version of cos(getPhi(lv)); see getPhi.

QEDbase.getCosTheta — Method

getCosTheta(lv)

Return the cosine of the theta angle for a given LorentzVectorLike.

Note

This is an equivalent but faster version of cos(getTheta(lv)); see getTheta.

QEDbase.getE — Method

getE(lv)

Return the energy component of a given LorentzVectorLike, i.e. its 0-component.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to E.

QEDbase.getEnergy — Function

Function alias for getE.

QEDbase.getGamma — Method

getGamma(lv)

Return the relativistic gamma factor for a given LorentzVectorLike, i.e. the inverse square root of one minus the beta vector squared.

Example

If (E,px,py,pz) is a LorentzVectorLike with beta vector β, this is equivalent to 1/sqrt(1- β^2).

QEDbase.getInvariantMass — Method

getInvariantMass(lv)

Return the the invariant mass of a given LorentzVectorLike, i.e. the square root of the minkowski dot with itself.

Example

If (t,x,y,z) is a LorentzVectorLike, this is equivalent to sqrt(t^2 - (x^2 + y^2 + z^2)).

Note

If the squared invariant mass m2 is negative, -sqrt(-m2) is returned.

QEDbase.getInvariantMass2 — Method

getInvariantMass2(lv)

Return the squared invariant mass of a given LorentzVectorLike, i.e. the minkowski dot with itself.

Example

If (t,x,y,z) is a LorentzVectorLike, this is equivalent to t^2 - (x^2 + y^2 + z^2).

QEDbase.getMag — Function

Functiom alias for getMagnitude.

QEDbase.getMag2 — Function

Functiom alias for getMagnitude2.

QEDbase.getMagnitude — Method

getMagnitude(lv)

Return the magnitude of a given LorentzVectorLike, i.e. the euklidian norm spatial components.

Example

If (t,x,y,z) is a LorentzVectorLike, this is equivalent to sqrt(x^2 + y^2 + z^2).

Warning

This function differs from a similar function for the TLorentzVector used in the famous ROOT library.

QEDbase.getMagnitude2 — Method

getMagnitude2(lv)

Return the square of the magnitude of a given LorentzVectorLike, i.e. the sum of the squared spatial components.

Example

If (t,x,y,z) is a LorentzVectorLike, this is equivalent to x^2+ y^2 + z^2.

Warning

This function differs from a similar function for the TLorentzVector used in the famous ROOT library.

QEDbase.getMass — Function

Function alias for getInvariantMass.

QEDbase.getMass2 — Function

Function alias for getInvariantMass2

QEDbase.getMinus — Method

getMinus(lv)

Return the minus component for a given LorentzVectorLike in light-cone coordinates.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to (E-pz)/2.

Note

The light-cone coordinates are defined w.r.t. to the 3-axis.

Warning

The definition `p^- := (E - p_z)/2 differs from the usual definition of light-cone coordinates in general relativity.

QEDbase.getMt — Function

Function alias for getTransverseMass

QEDbase.getMt2 — Function

Function alias for getTransverseMass2

QEDbase.getPerp — Function

Function alias for getTransverseMomentum.

QEDbase.getPerp2 — Function

Function alias for getTransverseMomentum2.

QEDbase.getPhi — Method

getPhi(lv)

Return the phi angle for a given LorentzVectorLike, i.e. the azimuthal angle of its spatial components in spherical coordinates.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to atan(py,px).

Note

The spherical coordinates are defined w.r.t. to the 3-axis.

QEDbase.getPlus — Method

getPlus(lv)

Return the plus component for a given LorentzVectorLike in light-cone coordinates.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to (E+pz)/2.

Note

The light-cone coordinates are defined w.r.t. to the 3-axis.

Warning

The definition `p^+ := (E + p_z)/2 differs from the usual definition of light-cone coordinates in general relativity.

QEDbase.getPt — Function

Function alias for getTransverseMomentum.

QEDbase.getPt2 — Function

Function alias for getTransverseMomentum2.

QEDbase.getPx — Method

getPx(lv)

Return the $p_x$ component of a given LorentzVectorLike, i.e. its 1-component.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to px.

QEDbase.getPy — Method

getPy(lv)

Return the $p_y$ component of a given LorentzVectorLike, i.e. its 2-component.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to py.

QEDbase.getPz — Method

getPz(lv)

Return the $p_z$ component of a given LorentzVectorLike, i.e. its 3-component.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to pz.

QEDbase.getRapidity — Method

getRapidity(lv)

Return the rapidity for a given LorentzVectorLike.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to 0.5*log((E+pz)/(E-pz)).

Note

The transverse components are defined w.r.t. to the 3-axis.

QEDbase.getRho — Function

Function alias for getMagnitude

QEDbase.getRho2 — Function

Function alias for getMagnitude2

QEDbase.getSinPhi — Method

getSinPhi(lv)

Return the sine of the phi angle for a given LorentzVectorLike.

Note

This is an equivalent but faster version of sin(getPhi(lv)); see getPhi.

QEDbase.getT — Function

getT(lv)

Return the 0-component of a given LorentzVectorLike.

Example

If (t,x,y,z) is a LorentzVectorLike, this is equivalent to t.

QEDbase.getTheta — Method

getTheta(lv)

Return the theta angle for a given LorentzVectorLike, i.e. the polar angle of its spatial components in spherical coordinates.

Example

If (E,px,py,pz) is a LorentzVectorLike with magnitude rho, this is equivalent to arccos(pz/rho), which is also equivalent to arctan(sqrt(px^2+py^2)/pz).

Note

The spherical coordinates are defined w.r.t. to the 3-axis.

QEDbase.getTransverseMass — Method

getTransverseMass(lv)

Return the transverse momentum for a given LorentzVectorLike, i.e. the square root of its squared transverse mass.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to sqrt(E^2 - pz^2).

Note

The transverse components are defined w.r.t. to the 3-axis.

Note

If the squared transverse mass mT2 is negative, -sqrt(-mT2) is returned.

QEDbase.getTransverseMass2 — Method

getTransverseMass2(lv)

Return the squared transverse mass for a given LorentzVectorLike, i.e. the difference of its squared 0- and 3-component.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to E^2 - pz^2.

Note

The transverse components are defined w.r.t. to the 3-axis.

QEDbase.getTransverseMomentum — Method

getTransverseMomentum(lv)

Return the transverse momentum for a given LorentzVectorLike, i.e. the magnitude of its transverse components.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to sqrt(px^2 + py^2).

Note

The transverse components are defined w.r.t. to the 3-axis.

QEDbase.getTransverseMomentum2 — Method

getTransverseMomentum2(lv)

Return the squared transverse momentum for a given LorentzVectorLike, i.e. the sum of its squared 1- and 2-component.

Example

If (E,px,py,pz) is a LorentzVectorLike, this is equivalent to px^2 + py^2.

Note

The transverse components are defined w.r.t. to the 3-axis.

QEDbase.getX — Function

getX(lv)

Return the 1-component of a given LorentzVectorLike.

Example

If (t,x,y,z) is a LorentzVectorLike, this is equivalent to x.

QEDbase.getY — Function

getY(lv)

Return the 2-component of a given LorentzVectorLike.

Example

If (t,x,y,z) is a LorentzVectorLike, this is equivalent to y.

QEDbase.getZ — Function

getZ(lv)

Return the 3-component of a given LorentzVectorLike.

Example

If (t,x,y,z) is a LorentzVectorLike, this is equivalent to z.

QEDbase.mdot — Function

Function alias for minkowski_dot.

QEDbase.minkowski_dot — Method

minkowski_dot(v1,v2)

Return the Minkowski dot product of two LorentzVectorLike.

Example

If (t1,x1,y1,z1) and (t2,x2,y2,z2) are two LorentzVectorLike, this is equivalent to

t1*t2 - (x1*x2 + y1*y2 + z1*z2)

Note

We use the mostly minus metric.

QEDbase.register_LorentzVectorLike — Method

register_LorentzVectorLike(T)

Function to register a custom type as a LorentzVectorLike.

Ensure the passed custom type has implemented at least the function getT, getX, getY, getZ and enables getter functions of the lorentz vector library for the given type. If additionally the functions setT!, setX!, setY!, setZ! are implemened for the passed custom type, also the setter functions of the Lorentz vector interface are enabled.

QEDbase.setCosTheta! — Method

setCosTheta!(lv,value)

Sets the cosine of the theta angle of a LorentzVectorLike to a given value.

Note

The value set with setCosTheta! is then returned by getCosTheta. Since the cosine of the theta angle is computed on the call of getCosTheta, the setter setCosTheta! changes several components of the given LorentzVectorLike.

QEDbase.setE! — Method

setE!(lv,value)

Sets the energy component of a given LorentzVectorLike to a given value.

Note

The value set with setE! is then returned by getE.

QEDbase.setEnergy! — Function

Function alias for setE!.

QEDbase.setMinus! — Method

setMinus!(lv,value)

Sets the minus component of a LorentzVectorLike to a given value.

Note

The value set with setMinus! is then returned by getMinus. Since the minus component is computed on the call of getMinus, the setter setMinus! changes several components of the given LorentzVectorLike.

QEDbase.setMt! — Function

Function alias for setTransverseMass!.

QEDbase.setPerp! — Function

Function alias for setTransverseMomentum!.

QEDbase.setPhi! — Method

setPhi!(lv,value)

Sets the phi angle of a LorentzVectorLike to a given value.

Note

The value set with setPhi! is then returned by getPhi. Since the phi angle is computed on the call of getPhi, the setter setPhi! changes several components of the given LorentzVectorLike.

QEDbase.setPlus! — Method

setPlus!(lv,value)

Sets the plus component of a LorentzVectorLike to a given value.

Note

The value set with setPlus! is then returned by getPlus. Since the plus component is computed on the call of getPlus, the setter setPlus! changes several components of the given LorentzVectorLike.

QEDbase.setPt! — Function

Function alias for setTransverseMomentum!.

QEDbase.setPx! — Method

setPx!(lv,value)

Sets the 1-component of a given LorentzVectorLike to a given value.

Note

The value set with setPx! is then returned by getPx.

QEDbase.setPy! — Method

setPy!(lv,value)

Sets the 2-component of a given LorentzVectorLike to a given value.

Note

The value set with setPy! is then returned by getPy.

QEDbase.setPz! — Method

setPz!(lv,value)

Sets the 3-component of a given LorentzVectorLike to a given value.

Note

The value set with setPz! is then returned by getPz.

QEDbase.setRapidity! — Method

setRapidity!(lv,value)

Sets the rapidity of a LorentzVectorLike to a given value.

Note

The value set with setRapidity! is then returned by getRapidity. Since the rapidity is computed on the call of setRapidity, the setter setRapidity! changes several components of the given LorentzVectorLike.

QEDbase.setRho! — Method

setRho!(lv,value)

Sets the magnitude of a LorentzVectorLike to a given value.

Note

The value set with setRho! is then returned by getRho. Since the magnitude is computed on the call of getRho, the setter setRho! changes several components of the given LorentzVectorLike.

QEDbase.setT! — Function

setT!(lv,value)

Sets the 0-component of a given LorentzVectorLike to the given value.

QEDbase.setTheta! — Method

setTheta!(lv,value)

Sets the theta angle of a LorentzVectorLike to a given value.

Note

The value set with setTheta! is then returned by getTheta. Since the theta angle is computed on the call of getTheta, the setter setTheta! changes several components of the given LorentzVectorLike.

QEDbase.setTransverseMass! — Method

setTransverseMass!(lv,value)

Sets the transverse mass of a LorentzVectorLike to a given value.

Note

The value set with setTransverseMass! is then returned by getTransverseMass. Since the transverse mass is computed on the call of getTransverseMass, the setter setTransverseMass! changes several components of the given LorentzVectorLike.

QEDbase.setTransverseMomentum! — Method

setTransverseMomentum!(lv,value)

Sets the transverse momentum of a LorentzVectorLike to a given value.

Note

The value set with setTransverseMomentum! is then returned by getTransverseMomentum. Since the transverse momentum is computed on the call of getTransverseMomentum, the setter setTransverseMomentum! changes several components of the given LorentzVectorLike.

QEDbase.setX! — Function

setX!(lv,value)

Sets the 1-component of a given LorentzVectorLike to the given value.

QEDbase.setY! — Function

setY!(lv,value)

Sets the 2-component of a given LorentzVectorLike to the given value.

QEDbase.setZ! — Function

setZ!(lv,value)

Sets the 3-component of a given LorentzVectorLike to the given value.

Lorentz Vectors

QEDbase.AbstractLorentzVector — Type

abstract type AbstractLorentzVector{T} <: StaticArraysCore.FieldVector{4, T}

Abstract type to model generic Lorentz vectors, i.e. elements of a minkowski-like space, where the component space is arbitray.

QEDbase.MLorentzVector — Type

mutable struct MLorentzVector{T} <: AbstractLorentzVector{T}

Concrete implementation of a generic mutable Lorentz vector. Each manipulation of an concrete implementation which is not self-contained (i.e. produces the same Lorentz vector type) will result in this type.

Fields

t::Any: t component

x::Any: x component

y::Any: y component

z::Any: z component

QEDbase.SLorentzVector — Type

struct SLorentzVector{T} <: AbstractLorentzVector{T}

Concrete implementation of a generic static Lorentz vector. Each manipulation of an concrete implementation which is not self-contained (i.e. produces the same Lorentz vector type) will result in this type.

Fields

t::Any: t component

x::Any: x component

y::Any: y component

z::Any: z component

QEDbase.mul — Method

mul(
    DM::Union{AbstractDiracMatrix, AbstractDiracVector},
    L::AbstractLorentzVector
) -> Any

Product of generic Lorentz vector with a Dirac tensor from the left. Basically, the multiplication is piped to the components from the Lorentz vector.

Multiplication operator

This also overloads the * operator for this types.

QEDbase.mul — Method

mul(
    L::AbstractLorentzVector,
    DM::Union{AbstractDiracMatrix, AbstractDiracVector}
) -> Any

Product of generic Lorentz vector with a Dirac tensor from the right. Basically, the multiplication is piped to the components from the Lorentz vector.

Multiplication operator

This also overloads the * operator for this types.

Four Momentum

QEDbase.MFourMomentum — Type

mutable struct MFourMomentum <: QEDbase.AbstractFourMomentum

Builds a mutable LorentzVector with real components used to statically model the four-momentum of a particle or field.

Fields

E::Float64: energy component

px::Float64: x component

py::Float64: y component

pz::Float64: z component

QEDbase.MFourMomentum — Method

MFourMomentum(t, x, y, z)

The interface transforms each number-like input to float64:

MFourMomentum(
    t::Union{Integer, Irrational, Rational},
    x::Union{Integer, Irrational, Rational},
    y::Union{Integer, Irrational, Rational},
    z::Union{Integer, Irrational, Rational}
) -> MFourMomentum

QEDbase.SFourMomentum — Type

struct SFourMomentum <: QEDbase.AbstractFourMomentum

Builds a static LorentzVectorLike with real components used to statically model the four-momentum of a particle or field.

Fields

E::Float64: energy component

px::Float64: x component

py::Float64: y component

pz::Float64: z component

QEDbase.SFourMomentum — Method

SFourMomentum(t, x, y, z)

The interface transforms each number-like input to float64:

SFourMomentum(
    t::Union{Integer, Irrational, Rational},
    x::Union{Integer, Irrational, Rational},
    y::Union{Integer, Irrational, Rational},
    z::Union{Integer, Irrational, Rational}
) -> SFourMomentum

QEDbase.assert_onshell — Method

assert_onshell(mom, mass)

Assertion if a FourMomentum mom is on-shell w.r.t a given mass mass.

See also

The precision of this functions is explained in isonshell.

QEDbase.isonshell — Method

isonshell(mom, mass)

On-shell check of a given four-momentum mom w.r.t. a given mass mass.

Precision

For AbstactFourMomentum, the element type is fixed to Float64, limiting the precision of comparisons between elements. The current implementation has been tested within the boundaries for energy scales E with 1e-9 <= E <= 1e5. In those bounds, the mass error, which is correctly detected as off-shell, is 1e-4 times the mean value of the components, but has at most the value 0.01, e.g. at the high energy end. The energy scales correspond to 0.5 meV for the lower bound and 50 GeV for the upper bound.

FourMomenta with real entries

if AbstractFourMomentum is updated to elementtypes T<:Real, the AbstractLorentzVector should be updated with the AbstractFourMomentum.

Dirac Tensors

QEDbase.AbstractDiracMatrix — Type

abstract type AbstractDiracMatrix{T} <: StaticArraysCore.FieldMatrix{4, 4, T}

Abstract type for Dirac matrices, i.e. matrix representations for linear mappings from a spinor space into another.

QEDbase.AbstractDiracVector — Type

abstract type AbstractDiracVector{T} <: StaticArraysCore.FieldVector{4, T}

Abstract type for Dirac vectors, e.g. four dimensional vectors from a spinor space.

QEDbase.AdjointBiSpinor — Type

struct AdjointBiSpinor <: AbstractDiracVector{ComplexF64}

Concrete type to model an adjoint Dirac four-spinor with complex-valued components. These are the elements of the dual spinor space.

QEDbase.BiSpinor — Type

struct BiSpinor <: AbstractDiracVector{ComplexF64}

Concrete type to model a Dirac four-spinor with complex-valued components. These are the elements of an actual spinor space.

QEDbase.DiracMatrix — Type

struct DiracMatrix <: AbstractDiracMatrix{ComplexF64}

Concrete type to model Dirac matrices, i.e. matrix representations of linear mappings between two spinor spaces.

QEDbase.mul — Method

mul(aBS::AdjointBiSpinor, BS::BiSpinor) -> ComplexF64

Tensor product of an adjoint with a standard bi-spinor resulting in a scalar.

Multiplication operator

This also overloads the * operator for this types.

QEDbase.mul — Method

mul(
    aBS::AdjointBiSpinor,
    DM::DiracMatrix,
    BS::BiSpinor
) -> ComplexF64

Tensor product of Dirac matrix sandwiched between an adjoint and a standard bi-spinor resulting in a scalar.

QEDbase.mul — Method

mul(
    aBS::AdjointBiSpinor,
    DM::DiracMatrix
) -> AdjointBiSpinor

Tensor product of an adjoint bi-spinor with a Dirac matrix resulting in another adjoint bi-spinor.

Multiplication operator

This also overloads the * operator for this types.

QEDbase.mul — Method

mul(BS::BiSpinor, aBS::AdjointBiSpinor) -> DiracMatrix

Tensor product of a standard with an adjoint bi-spinor resulting in a Dirac matrix.

Multiplication operator

This also overloads the * operator for this types.

QEDbase.mul — Method

mul(DM::DiracMatrix, BS::BiSpinor) -> BiSpinor

Tensor product of an Dirac matrix with a standard bi-spinor resulting in another standard bi-spinor.

Multiplication operator

This also overloads the * operator for this types.

QEDbase.mul — Method

mul(DM1::DiracMatrix, DM2::DiracMatrix) -> DiracMatrix

Tensor product two Dirac matrices resulting in another Dirac matrix.

Multiplication operator

This also overloads the * operator for this types.

Particles

QEDbase.Incoming — Type

Incoming <: ParticleDirection

Concrete implementation of a ParticleDirection to indicate that a particle is incoming in the context of a given process. Mostly used for dispatch.

julia> using QEDbase

julia> Incoming()
incoming

ParticleDirection Interface

Besides being a subtype of ParticleDirection, Incoming has

is_incoming(::Incoming) = true
is_outgoing(::Incoming) = false

QEDbase.Outgoing — Type

Outgoing <: ParticleDirection

Concrete implementation of a ParticleDirection to indicate that a particle is outgoing in the context of a given process. Mostly used for dispatch.

julia> using QEDbase

julia> Outgoing()
outgoing

ParticleDirection Interface

Besides being a subtype of ParticleDirection, Outgoing has

is_incoming(::Outgoing) = false
is_outgoing(::Outgoing) = true

QEDbase.ParticleDirection — Type

Abstract base type for the directions of particles in the context of processes, i.e. either they are incoming or outgoing. Subtypes of this are mostly used for dispatch.

QEDbase.AbstractDefinitePolarization — Type

Abstract base type for definite polarizations of BosonLike particles.

Concrete types are PolarizationX and PolarizationY.

QEDbase.AbstractDefiniteSpin — Type

Abstract base type for definite spins of FermionLike particles.

Concrete types are SpinUp and SpinDown.

QEDbase.AbstractIndefinitePolarization — Type

Abstract base type for indefinite polarizations of BosonLike particles.

One concrete type is AllPolarization.

QEDbase.AbstractIndefiniteSpin — Type

Abstract base type for indefinite spins of FermionLike particles.

One concrete type is AllSpin.

QEDbase.AbstractPolarization — Type

Abstract base type for the polarization of BosonLike particles.

QEDbase.AbstractSpin — Type

Abstract base type for the spin of FermionLike particles.

QEDbase.AbstractSpinOrPolarization — Type

Abstract base type for the spin or polarization of FermionLike or BosonLike particles, respectively.

QEDbase.AllPolarization — Type

Concrete type indicating that a BosonLike has an indefinite polarization and the differential cross section calculation should average or sum over all polarizations, depending on the direction (Incoming or Outgoing) of the particle in question.

julia> using QEDbase

julia> AllPol()
all polarizations

Alias

There is a built-in alias for AllPolarization:

julia> using QEDbase

julia> AllPolarization === AllPol
true

QEDbase.AllSpin — Type

Concrete type indicating that a FermionLike has an indefinite spin and the differential cross section calculation should average or sum over all spins, depending on the direction (Incoming or Outgoing) of the particle in question.

julia> using QEDbase

julia> AllSpin()
all spins

QEDbase.PolarizationX — Type

Concrete type which indicates, that a BosonLike has polarization in $x$-direction.

julia> using QEDbase

julia> PolX()
x-polarized

Coordinate axes

The notion of axes, e.g. $x$- and $y$-direction is just to distinguish two orthogonal polarization directions. However, if the three-momentum of the BosonLike is aligned to the $z$-axis of a coordinate system, the polarization axes define the $x$- or $y$-axis, respectively.

Alias

There is a built-in alias for PolarizationX:

julia> using QEDbase

julia> PolarizationX === PolX
true

QEDbase.PolarizationY — Type

Concrete type which indicates, that a BosonLike has polarization in $y$-direction.

julia> using QEDbase

julia> PolY()
y-polarized

Coordinate axes

The notion of axes, e.g. $x$- and $y$-direction is just to distinguish two orthogonal polarization directions. However, if the three-momentum of the BosonLike is aligned to the $z$-axis of a coordinate system, the polarization axes define the $x$- or $y$-axis, respectively.

Alias

There is a built-in alias for PolarizationY:

julia> using QEDbase

julia> PolarizationY === PolY
true

QEDbase.SpinDown — Type

Concrete type indicating that a FermionLike has spin-down.

julia> using QEDbase

julia> SpinDown()
spin down

QEDbase.SpinUp — Type

Concrete type indicating that a FermionLike has spin-up.

julia> using QEDbase

julia> SpinUp()
spin up

QEDbase._spin_index — Method

_spin_index(_::AbstractDefiniteSpin) -> Int64

Utility function, which converts subtypes of AbstractDefiniteSpin into an integer representation, e.g. an index:

SpinUp $\rightarrow 1$

SpinDown $\rightarrow 2$

This is useful in the implementation of some of base_state's overloads.

QEDbase._as_svec — Function

_as_svec(x)

Accepts a single object, an SVector of objects or a tuple of objects, and returns them in a single "layer" of SVector.

Intended for usage with base_state.

QEDbase.base_state — Function

    base_state(
        particle::AbstractParticleType,
        direction::ParticleDirection,
        momentum::QEDbase.AbstractFourMomentum,
        [spin_or_pol::AbstractSpinOrPolarization]
    )

Return the base state of a directed on-shell particle with a given four-momentum. For internal usage only.

Input

particle – the type of the particle, e.g. Electron, Positron, or Photon.

direction – the direction of the particle, i.e. Incoming or Outgoing.

momentum – the four-momentum of the particle

[spin_or_pol] – if given, the spin or polarization of the particle, e.g. SpinUp/SpinDown or PolarizationX/PolarizationY.

Output

The output type of base_state depends on wether the spin or polarization of the particle passed in is specified or not.

If spin_or_pol is passed, the output of base_state is

base_state(::Fermion,     ::Incoming, mom, spin_or_pol) # -> BiSpinor
base_state(::AntiFermion, ::Incoming, mom, spin_or_pol) # -> AdjointBiSpinor
base_state(::Fermion,     ::Outgoing, mom, spin_or_pol) # -> AdjointBiSpinor
base_state(::AntiFermion, ::Outgoing, mom, spin_or_pol) # -> BiSpinor
base_state(::Photon,      ::Incoming, mom, spin_or_pol) # -> SLorentzVector{ComplexF64}
base_state(::Photon,      ::Outgoing, mom, spin_or_pol) # -> SLorentzVector{ComplexF64}

If spin_or_pol is of type AllPolarization or AllSpin, the output is an SVector with both spin/polarization alignments:

base_state(::Fermion,     ::Incoming, mom) # -> SVector{2,BiSpinor}
base_state(::AntiFermion, ::Incoming, mom) # -> SVector{2,AdjointBiSpinor}
base_state(::Fermion,     ::Outgoing, mom) # -> SVector{2,AdjointBiSpinor}
base_state(::AntiFermion, ::Outgoing, mom) # -> SVector{2,BiSpinor}
base_state(::Photon,      ::Incoming, mom) # -> SVector{2,SLorentzVector{ComplexF64}}
base_state(::Photon,      ::Outgoing, mom) # -> SVector{2,SLorentzVector{ComplexF64}}

Example

using QEDbase

mass = 1.0                              # set electron mass to 1.0
px,py,pz = rand(3)                      # generate random spatial components
E = sqrt(px^2 + py^2 + pz^2 + mass^2)   # compute energy, i.e. the electron is on-shell
mom = SFourMomentum(E, px, py, pz)      # initialize the four-momentum of the electron

# compute the state of an incoming electron with spin = SpinUp
# note: base_state is not exported!
electron_state = base_state(Electron(), Incoming(), mom, SpinUp())

julia> using QEDbase

julia> mass = 1.0; px,py,pz = (0.1, 0.2, 0.3); E = sqrt(px^2 + py^2 + pz^2 + mass^2); mom = SFourMomentum(E, px, py, pz)
4-element SFourMomentum with indices SOneTo(4):
 1.0677078252031311
 0.1
 0.2
 0.3

julia> electron_state = base_state(Electron(), Incoming(), mom, SpinUp())
4-element BiSpinor with indices SOneTo(4):
   1.4379526505428235 + 0.0im
                  0.0 + 0.0im
 -0.20862995724285552 + 0.0im
 -0.06954331908095185 - 0.1390866381619037im

julia> electron_states = base_state(Electron(), Incoming(), mom, AllSpin())
2-element StaticArraysCore.SVector{2, BiSpinor} with indices SOneTo(2):
 [1.4379526505428235 + 0.0im, 0.0 + 0.0im, -0.20862995724285552 + 0.0im, -0.06954331908095185 - 0.1390866381619037im]
 [0.0 + 0.0im, 1.4379526505428235 + 0.0im, -0.06954331908095185 + 0.1390866381619037im, 0.20862995724285552 + 0.0im]

Iterator convenience

The returned objects of base_state can be consistently wrapped in an SVector for iteration using _as_svec.

This way, a loop like the following becomes possible when spin may be definite or indefinite.

for state in QEDbase._as_svec(base_state(Electron(), Incoming(), momentum, spin))
    # ...
end

Conventions

For an incoming fermion with momentum $p$, we use the explicit formula:

\[u_\sigma(p) = \frac{\gamma^\mu p_\mu + m}{\sqrt{\vert p_0\vert + m}} \eta_\sigma,\]

where the elementary base spinors are given as

\[\eta_1 = (1, 0, 0, 0)^T\\ \eta_2 = (0, 1, 0, 0)^T\]

For an outgoing anti-fermion with momentum $p$, we use the explicit formula:

\[v_\sigma(p) = \frac{-\gamma^\mu p_\mu + m}{\sqrt{\vert p_0\vert + m}} \chi_\sigma,\]

where the elementary base spinors are given as

\[\chi_1 = (0, 0, 1, 0)^T\\ \chi_2 = (0, 0, 0, 1)^T\]

For outgoing fermions and incoming anti-fermions with momentum $p$, the base state is given as the Dirac-adjoint of the respective incoming fermion or outgoing anti-fermion state:

\[\overline{u}_\sigma(p) = u_\sigma^\dagger \gamma^0\\ \overline{v}_\sigma(p) = v_\sigma^\dagger \gamma^0\]

where $v_\sigma$ is the base state of the respective outgoing anti-fermion.

For a photon with four-momentum $k^\mu = \omega (1, \cos\phi \sin\theta, \sin\phi \sin\theta, \cos\theta)$, the two polarization vectors are given as

\[\begin{align*} \epsilon^\mu_1 &= (0, \cos\theta \cos\phi, \cos\theta \sin\phi, -\sin\theta),\\ \epsilon^\mu_2 &= (0, -\sin\phi, \cos\phi, 0). \end{align*}\]

Warning

In the current implementation there are no checks built-in, which verify the passed arguments whether they describe on-shell particles, i.e. p*p≈mass^2. Using base_state with off-shell particles will cause unpredictable behavior.

QEDbase.AbstractParticleType — Type

AbstractParticleType <: AbstractParticle

This is the abstract base type for every species of particles. All functionalities defined on subtypes of AbstractParticleType should be static, i.e. known at compile time. For adding runtime information, e.g. four-momenta or particle states, to a particle, consider implementing a concrete subtype of AbstractParticle instead, which may have a type parameter P<:AbstractParticleType.

Concrete built-in subtypes of AbstractParticleType are Electron, Positron and Photon.

QEDbase.AntiBoson — Type

Abstract base type for anti-bosons as distinct from its particle counterpart Boson.

particle interface

All subtypes of AntiBoson have

is_boson(::AntiBoson) = true
is_particle(::AntiBoson) = false
is_anti_particle(::AntiBoson) = true

QEDbase.AntiFermion — Type

Abstract base type for anti-fermions as distinct from its particle counterpart Fermion.

particle interface

All subtypes of AntiFermion have

is_fermion(::AntiFermion) = true
is_particle(::AntiFermion) = false
is_anti_particle(::AntiFermion) = true

QEDbase.Boson — Type

Abstract base type for bosons as distinct from its anti-particle counterpart AntiBoson.

particle interface

All subtypes of Boson have

is_boson(::Boson) = true
is_particle(::Boson) = true
is_anti_particle(::Boson) = false

QEDbase.BosonLike — Type

Abstract base types for particle species that act like bosons in the sense of particle statistics.

particle interface

Every concrete subtype of BosonLike has is_boson(::BosonLike) = true.

QEDbase.Electron — Type

Concrete type for electrons as a particle species. Mostly used for dispatch.

julia> using QEDbase

julia> Electron()
electron

particle interface

Besides being a subtype of Fermion, objects of type Electron have

mass(::Electron) = 1.0
charge(::Electron) = -1.0

QEDbase.Fermion — Type

Abstract base type for fermions as distinct from AntiFermions.

particle interface

All subtypes of Fermion have

is_fermion(::Fermion) = true
is_particle(::Fermion) = true
is_anti_particle(::Fermion) = false

QEDbase.FermionLike — Type

Abstract base types for particle species that act like fermions in the sense of particle statistics.

particle interface

Every concrete subtype of FermionLike has is_fermion(::FermionLike) = true.

QEDbase.MajoranaBoson — Type

Abstract base type for majorana-bosons, i.e. bosons which are their own anti-particles.

particle interface

All subtypes of MajoranaBoson have

is_boson(::MajoranaBoson) = true
is_particle(::MajoranaBoson) = true
is_anti_particle(::MajoranaBoson) = true

QEDbase.MajoranaFermion — Type

Abstract base type for majorana-fermions, i.e. fermions which are their own anti-particles.

particle interface

All subtypes of MajoranaFermion have

is_fermion(::MajoranaFermion) = true
is_particle(::MajoranaFermion) = true
is_anti_particle(::MajoranaFermion) = true

QEDbase.Photon — Type

Concrete type for the photons as a particle species. Mostly used for dispatch.

julia> using QEDbase

julia> Photon()
photon

particle interface

Besides being a subtype of MajoranaBoson, Photon has

mass(::Photon) = 0.0
charge(::Photon) = 0.0

QEDbase.Positron — Type

Concrete type for positrons as a particle species. Mostly used for dispatch.

julia> using QEDbase

julia> Positron()
positron

particle interface

Besides being a subtype of AntiFermion, objects of type Positron have

mass(::Positron) = 1.0
charge(::Positron) = 1.0

QEDbase.AbstractParticle — Type

Abstract base type for every type which might be considered a particle in the context of QED.jl. For every (concrete) subtype of AbstractParticle, there are two kinds of interface functions implemented: static functions and property functions. The static functions provide information on what kind of particle it is (defaults are written in square brackets)

    is_fermion(::AbstractParticle)::Bool [= false]
    is_boson(::AbstractParticle)::Bool [= false]
    is_particle(::AbstractParticle)::Bool [= true]
    is_anti_particle(::AbstractParticle)::Bool [= false]

If the output of those functions differ from the defaults for a subtype of AbstractParticle, these functions need to be overwritten. The second type of functions define a hard interface for AbstractParticle:

    mass(::AbstractParticle)::Real
    charge(::AbstractParticle)::Real

These functions must be implemented in order to have the subtype of AbstractParticle work with the functionalities of QEDprocesses.jl.

QEDbase.charge — Function

charge(::AbstractParticle)::Real

Interface function for particles. Return the electric charge of a particle (in units of the elementary electric charge).

This needs to be implemented for each concrete subtype of AbstractParticle.

QEDbase.is_anti_particle — Method

is_anti_particle(_::AbstractParticle) -> Bool

Interface function for particles. Return true if the passed subtype of AbstractParticle can be considered an anti particle as distinct from their particle counterpart, and false otherwise. The default implementation of is_anti_particle for every subtype of AbstractParticle will always return false.

QEDbase.is_boson — Method

is_boson(_::AbstractParticle) -> Bool

Interface function for particles. Return true if the passed subtype of AbstractParticle can be considered a boson in the sense of particle statistics, and false otherwise. The default implementation of is_boson for every subtype of AbstractParticle will always return false.

QEDbase.is_fermion — Method

is_fermion(_::AbstractParticle) -> Bool

Interface function for particles. Return true if the passed subtype of AbstractParticle can be considered a fermion in the sense of particle statistics, and false otherwise. The default implementation of is_fermion for every subtype of AbstractParticle will always return false.

QEDbase.is_particle — Method

is_particle(_::AbstractParticle) -> Bool

Interface function for particles. Return true if the passed subtype of AbstractParticle can be considered a particle as distinct from anti-particles, and false otherwise. The default implementation of is_particle for every subtype of AbstractParticle will always return true.

QEDbase.mass — Function

mass(particle::AbstractParticle)::Real

Interface function for particles. Return the rest mass of a particle (in units of the electron mass).

This needs to be implemented for each concrete subtype of AbstractParticle.