Phase Space Description

Frames and Coordinate Systems

QEDbase.AbstractCoordinateSystem — Type

AbstractCoordinateSystem

TBW

QEDbase.AbstractFrameOfReference — Type

AbstractFrameOfReference

TBW

QEDbase.AbstractPhasespaceDefinition — Type

AbstractPhasespaceDefinition

TBW

Stateful Particles

QEDbase.AbstractParticleStateful — Type

AbstractParticleStateful <: QEDbase.AbstractParticle

Abstract base type for the representation of a particle with a state. It requires the following interface functions to be provided:

particle_direction: Returning the particle's direction.

particle_species: Returning the particle's species.

momentum: Returning the particle's momentum.

Implementations for is_fermion, is_boson, is_particle, is_anti_particle, is_incoming, is_outgoing, mass, and charge are automatically provided using the interface functions above to fulfill the QEDbase.AbstractParticle interface.

QEDbase.momentum — Function

momentum(part::AbstractParticleStateful)

Interface function that must return the particle's AbstractFourMomentum.

Phasespace Points

Types

QEDbase.AbstractPhaseSpacePoint — Type

AbstractPhaseSpacePoint{PROC, MODEL, PSDEF, IN_PARTICLES, OUT_PARTICLES}

Representation of a point in the phase space of a process. It has several template arguments:

PROC <:AbstractProcessDefinition

MODEL <:AbstractModelDefinition

PSDEF <:AbstractPhasespaceDefinition

IN_PARTICLES <:Tuple{Vararg{AbstractParticleStateful}}: The tuple type of all the incoming [AbstractParticleStateful`](@ref)s.

OUT_PARTICLES <:Tuple{Vararg{AbstractParticleStateful}}: The tuple type of all the outgoing [AbstractParticleStateful`](@ref)s.

The following interface functions must be provided:

Base.getindex(psp::AbstractPhaseSpacePoint, dir::ParticleDirection, n::Int): Return the nth AbstractParticleStateful of the given direction. Throw BoundsError for invalid indices.

particles(psp::AbstractPhaseSpacePoint, dir::ParticleDirection): Return the particle tuple (type IN_PARTICLES or OUT_PARTICLES depending on dir)

process: Return the process.

model: Return the model.

phase_space_definition: Return the phase space definition.

From this, the following functions are automatically derived:

momentum(psp::AbstractPhaseSpacePoint, dir::ParticleDirection, n::Int): Return the momentum of the nth AbstractParticleStateful of the given direction.

momenta(psp::PhaseSpacePoint, ::ParticleDirection): Return a Tuple of all the momenta of the given direction.

Furthermore, an implementation of an AbstractPhaseSpacePoint has to verify on construction that it is valid, i.e., the following conditions are fulfilled:

IN_PARTICLES must match incoming_particles(::PROC) in length, order, and type or be an empty Tuple.

OUT_PARTICLES must match the outgoing_particles(::PROC) in length, order, and type, or be an empty Tuple.

IN_PARTICLES and OUT_PARTICLES may not both be empty.

If IN_PARTICLES is non-empty, AbstractPhaseSpacePoint <: AbstractInPhaseSpacePoint is true. Likewise, if OUT_PARTICLES is non-empty, AbstractPhaseSpacePoint <: AbstractOutPhaseSpacePoint is true. Consequently, if both IN_PARTICLES and OUT_PARTICLES are non-empty, both <: statements are true.

QEDbase.AbstractInPhaseSpacePoint — Type

AbstractInPhaseSpacePoint

A partial type specialization on AbstractPhaseSpacePoint which can be used for dispatch in functions requiring only the in channel of the phase space to exist, for example implementations of _incident_flux. No restrictions are imposed on the out-channel, which may or may not exist.

See also: AbstractOutPhaseSpacePoint

QEDbase.AbstractOutPhaseSpacePoint — Type

AbstractOutPhaseSpacePoint

A partial type specialization on AbstractPhaseSpacePoint which can be used for dispatch in functions requiring only the out channel of the phase space to exist. No restrictions are imposed on the in-channel, which may or may not exist.

See also: AbstractInPhaseSpacePoint

Interface

QEDbase.process — Function

process(psp::AbstractPhaseSpacePoint)

Return the phase space point's set process.

QEDbase.model — Function

model(psp::AbstractPhaseSpacePoint)

Return the phase space point's set model.

QEDbase.phase_space_definition — Function

phase_space_definition(psp::AbstractPhaseSpacePoint)

Return the phase space point's set phase space definition.

Convenience Functions

QEDbase.particle_direction — Function

particle_direction(part::AbstractParticleStateful)

Interface function that must return the particle's ParticleDirection.

QEDbase.particle_species — Function

particle_species(part::AbstractParticleStateful)

Interface function that must return the particle's AbstractParticleType, e.g. QEDcore.Electron().

QEDbase.momenta — Function

momenta(psp::AbstractPhaseSpacePoint, ::ParticleDirection)

Return a Tuple of all the particles' momenta for the given ParticleDirection.