Utility

Errors

QEDbase.InvalidInputError — Type

InvalidInputError(msg::String)

Exception which is thrown if a function input is invalid.

QEDbase.RegistryError — Type

struct RegistryError <: Exception

Exception raised, if a certain type target_type can not be registed for a certain interface, since there needs the function func to be impleemnted.

Fields

func::Function

target_type::Any

Helper functions

QEDbase._as_svec — Function

_as_svec(x)

Accepts a single object, an SVector of objects or a tuple of objects, and returns them in a single "layer" of SVector.

Useful with base_state.