Probability and Cross Section

Interface

QEDbase._incident_flux — Function

_incident_flux(in_psp::InPhaseSpacePoint{PROC,MODEL}) where {
    PROC <: AbstractProcessDefinition,
    MODEL <: AbstractModelDefinition,
}

Interface function which returns the incident flux of the given scattering process for a given InPhaseSpacePoint.

Differential and total probability

QEDbase.differential_probability — Function

differential_probability(phase_space_point::AbstractPhaseSpacePoint)

If the given phase spaces are physical, return differential probability evaluated on a phase space point. Zero otherwise.

QEDbase.unsafe_differential_probability — Function

unsafe_differential_probability(phase_space_point::AbstractPhaseSpacePoint)

Return differential probability evaluated on a phase space point without checking if the given phase space(s) are physical.

QEDbase.total_probability — Function

total_probability(in_psp::AbstractInPhaseSpacePoint)

Return the total probability of a given AbstractInPhaseSpacePoint.

Differential and total cross section

QEDbase.differential_cross_section — Function

differential_cross_section(phase_space_point::PhaseSpacePoint)

If the given phase spaces are physical, return differential cross section evaluated on a phase space point. Zero otherwise.

QEDbase.unsafe_differential_cross_section — Function

unsafe_differential_cross_section(phase_space_point::AbstractPhaseSpacePoint)

Return the differential cross section evaluated on a phase space point without checking if the given phase space is physical.

QEDbase.total_cross_section — Function

total_cross_section(in_psp::AbstractInPhaseSpacePoint)

Return the total cross section for a given AbstractInPhaseSpacePoint.